refactor(shared): extract voice command parser from VoiceService
VoiceService was 258 lines mixing three orthogonal concerns:
TTS playback, STT session management, and command-string parsing.
The parser is the easiest peel: 25 lines of pure regex matching
with no Angular dependencies and 16 dedicated tests already.

Move COMMAND_PATTERNS + the parser into voice-command-parser.ts as
parseVoiceCommand(). VoiceService imports it; the static
VoiceService.parseCommand becomes a free function, and the
VoiceCommand type now lives next to its parser. Public re-export
keeps the same `from '@yumney/shared/models'` import path for
consumers.

- voice.service.ts: 258 -> 219 lines
