+++ client/libs/shared/models/src/lib/voice.service.spec.ts
@@ -2,70 +2,71 @@ import { signal } from '@angular/core';
 import { TestBed } from '@angular/core/testing';
 import { UserPreferencesService } from './user-preferences.service';
 import { VoiceService } from './voice.service';
+import { parseVoiceCommand } from './voice-command-parser';
 
-describe('VoiceService.parseCommand', () => {
+describe('parseVoiceCommand', () => {
   it('should parse "next step" as next command', () => {
-    expect(VoiceService.parseCommand('next step')).toEqual({ type: 'next' });
+    expect(parseVoiceCommand('next step')).toEqual({ type: 'next' });
   });
 
   it('should parse German "nächster schritt" as next command', () => {
-    expect(VoiceService.parseCommand('nächster schritt')).toEqual({ type: 'next' });
+    expect(parseVoiceCommand('nächster schritt')).toEqual({ type: 'next' });
   });
 
   it('should parse "previous step" as previous command', () => {
-    expect(VoiceService.parseCommand('previous step')).toEqual({ type: 'previous' });
+    expect(parseVoiceCommand('previous step')).toEqual({ type: 'previous' });
   });
 
   it('should parse German "vorheriger schritt" as previous command', () => {
-    expect(VoiceService.parseCommand('vorheriger schritt')).toEqual({ type: 'previous' });
+    expect(parseVoiceCommand('vorheriger schritt')).toEqual({ type: 'previous' });
   });
 
   it('should parse "repeat" as repeat command', () => {
-    expect(VoiceService.parseCommand('repeat')).toEqual({ type: 'repeat' });
+    expect(parseVoiceCommand('repeat')).toEqual({ type: 'repeat' });
   });
 
   it('should parse "stop" as stop command', () => {
-    expect(VoiceService.parseCommand('stop')).toEqual({ type: 'stop' });
+    expect(parseVoiceCommand('stop')).toEqual({ type: 'stop' });
   });
 
   it('should parse German "stopp" as stop command', () => {
-    expect(VoiceService.parseCommand('stopp')).toEqual({ type: 'stop' });
+    expect(parseVoiceCommand('stopp')).toEqual({ type: 'stop' });
   });
 
   it('should parse "ingredients" as ingredients command', () => {
-    expect(VoiceService.parseCommand('ingredients')).toEqual({ type: 'ingredients' });
+    expect(parseVoiceCommand('ingredients')).toEqual({ type: 'ingredients' });
   });
 
   it('should parse German "zutaten" as ingredients command', () => {
-    expect(VoiceService.parseCommand('zutaten')).toEqual({ type: 'ingredients' });
+    expect(parseVoiceCommand('zutaten')).toEqual({ type: 'ingredients' });
   });
 
   it('should parse "timer 5 minutes" as timer command with minutes', () => {
-    expect(VoiceService.parseCommand('timer 5 minutes')).toEqual({ type: 'timer', minutes: 5 });
+    expect(parseVoiceCommand('timer 5 minutes')).toEqual({ type: 'timer', minutes: 5 });
   });
 
   it('should parse German "timer 10 minuten" as timer command with minutes', () => {
-    expect(VoiceService.parseCommand('timer 10 minuten')).toEqual({ type: 'timer', minutes: 10 });
+    expect(parseVoiceCommand('timer 10 minuten')).toEqual({ type: 'timer', minutes: 10 });
   });
 
   it('should parse "5 minute timer" as timer command', () => {
-    expect(VoiceService.parseCommand('5 minute timer')).toEqual({ type: 'timer', minutes: 5 });
+    expect(parseVoiceCommand('5 minute timer')).toEqual({ type: 'timer', minutes: 5 });
   });
 
   it('should be case-insensitive', () => {
-    expect(VoiceService.parseCommand('NEXT STEP')).toEqual({ type: 'next' });
+    expect(parseVoiceCommand('NEXT STEP')).toEqual({ type: 'next' });
   });
 
   it('should ignore surrounding whitespace', () => {
-    expect(VoiceService.parseCommand('  next step  ')).toEqual({ type: 'next' });
+    expect(parseVoiceCommand('  next step  ')).toEqual({ type: 'next' });
   });
 
   it('should return null for unknown phrases', () => {
-    expect(VoiceService.parseCommand('do something random')).toBeNull();
+    expect(parseVoiceCommand('do something random')).toBeNull();
   });
 
   it('should return null for empty input', () => {
-    expect(VoiceService.parseCommand('')).toBeNull();
+    expect(parseVoiceCommand('')).toBeNull();
   });
 });
