+++ client/libs/shared/models/src/lib/voice.service.ts
@@ -1,35 +1,8 @@
 import { Injectable, inject, signal } from '@angular/core';
 import { UserPreferencesService } from './user-preferences.service';
+import { parseVoiceCommand, type VoiceCommand } from './voice-command-parser';
 
-export type VoiceCommand =
-  | { type: 'next' }
-  | { type: 'previous' }
-  | { type: 'repeat' }
-  | { type: 'stop' }
-  | { type: 'ingredients' }
-  | { type: 'timer'; minutes: number };
-
-const COMMAND_PATTERNS: Array<{
-  match: RegExp;
-  build: (groups: RegExpMatchArray) => VoiceCommand;
-}> = [
-  { match: /^(next step|next|nächster schritt|weiter)$/i, build: () => ({ type: 'next' }) },
-  {
-    match: /^(previous step|previous|back|vorheriger schritt|zurück)$/i,
-    build: () => ({ type: 'previous' }),
-  },
-  { match: /^(repeat|wiederhole|wiederholen)$/i, build: () => ({ type: 'repeat' }) },
-  { match: /^(stop|stopp|stop it|halt)$/i, build: () => ({ type: 'stop' }) },
-  { match: /^(ingredients|zutaten)$/i, build: () => ({ type: 'ingredients' }) },
-  {
-    match: /^timer\s+(\d{1,3})\s*(?:minutes?|min|minuten|minute)$/i,
-    build: (match) => ({ type: 'timer', minutes: Number(match[1]) }),
-  },
-  {
-    match: /^(\d{1,3})\s*(?:minute|minuten|min)\s+timer$/i,
-    build: (match) => ({ type: 'timer', minutes: Number(match[1]) }),
-  },
-];
+export { type VoiceCommand };
 
 interface SpeechRecognitionLike {
   lang: string;
@@ -111,7 +84,7 @@ export class VoiceService {
     this.beginSession({
       continuous: true,
       onTranscript: (transcript) => {
-        const command = VoiceService.parseCommand(transcript);
+        const command = parseVoiceCommand(transcript);
         if (command) onCommand(command);
       },
     });
@@ -142,7 +115,7 @@ export class VoiceService {
       continuous: true,
       onTranscript: (transcript) => {
         if (transcript === '') return;
-        const command = VoiceService.parseCommand(transcript);
+        const command = parseVoiceCommand(transcript);
         if (command) {
           onCommand(command);
           return;
@@ -226,18 +199,6 @@ export class VoiceService {
     this.recognition = recognition;
     recognition.start();
     this.isListening.set(true);
-  }
-
-  static parseCommand(transcript: string): VoiceCommand | null {
-    const normalized = transcript.trim().toLowerCase();
-    if (normalized === '') return null;
-    for (const pattern of COMMAND_PATTERNS) {
-      const match = normalized.match(pattern.match);
-      if (match) {
-        return pattern.build(match);
-      }
-    }
-    return null;
   }
 
   private createRecognition(): SpeechRecognitionLike | null {
